Create Sanjay and Rahul
Sanjay and Rahul have a set of color crayons. Sanjay has a set of colors whereas Rahul has a set of colors. Write a program to find the colors they have in common. Print the colors that Rahul has but not Sanjay. Sanjay lost his Green color. Update this information.  

Note: Sort the sets before printing