Create concatenating three dictionaries
Write a program to concatenate three dictionaries to create a new one. 

Input format
The first line of the input consists of the value of n.

The next n lines are the dict1 elements.

The next n lines are the dict2 elements.

The next n lines are the dict3 elements.

Output format
The output prints the concatenated dictionaries.